Fix issue when locking and unlocking in overview causes TypeError
Non-initialized _windowOverlayIconsExtension property was accessed
because disable() uninitialized the property and enable() cannot
easily initialize it back.

If we wanted to initialize the property in enable() we would need to
find all instances of WindowOverlay and initialize them.
It is simpler to just leave the initialization to the instance's _init()
method. However this causes a strange behavior that when you go to
window overview, lock the screen and unlock it, the window icons are not
shown until you reopen the overview. But I can live with this.

See https://github.com/sustmi/gnome-shell-extensions-sustmi/issues/44
for details.