Fix stacktraces caused by double-destroying St.Bin actor
The St.Bin actor is children of a parentActor and gets destroyed when
the parent is destroyed. When this extension tried to destroy the
actor in disable() function GNOME Shell issued a warning.

In order to fix this double-destroy issue we connect to the "destroy"
signal and check whether the actor was already destroyed.

Fixes: https://github.com/sustmi/gnome-shell-extension-windowoverlay-icons/issues/4